This is my first open source web application. …
This project is solely developed to apply my learning and my testing on applying different architecture. I am implementing the latest technologies of the current which I am learning.
Features of this project is not so rich but the developers can be benefited with this project and get help of the technologies.